+++ artemis-test/src/main/java/com/artemis/system/OptimizedSystemAdditional.java
@@ -7,7 +7,7 @@ import com.artemis.systems.EntityProcessingSystem;
 public class OptimizedSystemAdditional extends EntityProcessingSystem {
 
 	public OptimizedSystemAdditional() {
-		super(null);
+		super(Aspect.all());
 
 		setEnabled(true);
 		begin();
@@ -24,7 +24,5 @@ public class OptimizedSystemAdditional extends EntityProcessingSystem {
 	}
 
 	@Override
-	protected void process(Entity e) {
-	}
-
+	protected void process(Entity e) {}
 }