+++ artemis/src/main/java/com/artemis/AspectSubscriptionManager.java
@@ -1,0 +1,50 @@
+package com.artemis;
+
+import com.artemis.utils.Bag;
+
+import java.util.BitSet;
+import java.util.HashMap;
+import java.util.Map;
+
+public class AspectSubscriptionManager extends Manager {
+
+	private final Map<Aspect.Builder, EntitySubscription> subscriptionMap;
+	private Bag<EntitySubscription> subscriptions;
+
+	public AspectSubscriptionManager() {
+		subscriptionMap = new HashMap<Aspect.Builder, EntitySubscription>();
+		subscriptions = new Bag<EntitySubscription>();
+	}
+
+	public EntitySubscription get(Aspect.Builder builder) {
+		EntitySubscription subscription = subscriptionMap.get(builder);
+		return (subscription != null) ? subscription : createSubscription(builder);
+	}
+
+	private EntitySubscription createSubscription(Aspect.Builder builder) {
+		EntitySubscription entitySubscription = new EntitySubscription(world, builder);
+		subscriptionMap.put(builder, entitySubscription);
+		subscriptions.add(entitySubscription);
+		return entitySubscription;
+	}
+
+	void process(WildBag<Entity> added, WildBag<Entity> changed, WildBag<Entity> deleted) {
+		Object[] subscribers = subscriptions.getData();
+		for (int i = 0, s = subscriptions.size(); s > i; i++) {
+			EntitySubscription subscriber = (EntitySubscription)subscribers[i];
+			subscriber.process(added, changed, deleted);
+		}
+		added.setSize(0);
+		changed.setSize(0);
+		deleted.setSize(0);
+	}
+
+	// TODO: heed later added subscriptions too
+	void processComponentIdentity(int id, BitSet componentBits) {
+		Object[] subscribers = subscriptions.getData();
+		for (int i = 0, s = subscriptions.size(); s > i; i++) {
+			EntitySubscription subscriber = (EntitySubscription)subscribers[i];
+			subscriber.processComponentIdentity(id, componentBits);
+		}
+	}
+}