+++ artemis/src/main/java/com/artemis/EntitySystem.java
@@ -1,7 +1,5 @@
 package com.artemis;
 
-import java.util.BitSet;
-
 import com.artemis.utils.Bag;
 import com.artemis.utils.ImmutableBag;
 import com.artemis.utils.IntBag;
@@ -16,29 +14,12 @@ import com.artemis.utils.IntBag;
  *
  * @author Arni Arent
  */
-public abstract class EntitySystem implements EntityObserver {
+public abstract class EntitySystem extends BaseSystem
+		implements EntitySubscription.SubscriptionListener {
 
-	/** The world this system belongs to. */
-	protected World world;
+	private final Aspect.Builder aspectConfiguration;
+	protected IntBag actives;
 	protected Entity flyweight;
-	
-	/* 
-	 * actives = only contains entities, typically sorted ASC by entity.id 
-	 * activesIsDirty = indicates that actives isn't sorted; needs rebuilding 
-	 */
-	private IntBag actives;
-	private final BitSet activeIds;
-	private boolean activesIsDirty;
-	
-	/** If the system is passive or not. */
-	private boolean passive;
-	/** If the system is enabled or not. */
-	private boolean enabled;
-	/** If the system is interested in no entities at all. */
-	private boolean dummy;
-	private Aspect.Builder aspectConfiguration;
-	private Aspect aspect;
-	private final BitSet aspectCache = new BitSet();
 
 	/**
 	 * Creates an entity system that uses the specified aspect as a matcher
@@ -48,85 +29,39 @@ public abstract class EntitySystem implements EntityObserver {
 	 *			to match against entities
 	 */
 	public EntitySystem(Aspect.Builder aspect) {
-		this.aspectConfiguration = aspect;
-		activeIds = new BitSet();
-		actives = new IntBag();
-		
-		enabled = true;
+		super();
+		if (aspect == null) {
+			String error = "Aspect.Builder was null; to use systems which " +
+					"do not subscribe to entities, extend VoidEntitySystem or " +
+					"BaseSystem directly.";
+			throw new NullPointerException(error);
+		}
+
+		aspectConfiguration = aspect;
 	}
 
+	protected void setWorld(World world) {
+		super.setWorld(world);
 
-	/**
-	 * Called before processing of entities begins.
-	 * <p>
-	 * <b>Nota Bene:</b> Any entities created in this method
-	 * won't become active until the next system starts processing
-	 * or when a new processing rounds beings, whichever comes first.
-	 * </p>
-	 */
-	protected void begin() {}
+		EntitySubscription subscription = getSubscription();
+		subscription.addSubscriptionListener(this);
+		actives = subscription.getEntities();
 
-	/**
-	 * Process all entities this system is interested in.
-	 */
-	public final void process() {
-		if(enabled && checkProcessing()) {
-			begin();
-			
-			if (activesIsDirty) // && world.isRebuildingIndexAllowed())
-				rebuildCompressedActives();
-			
-			processEntities(actives);
+		flyweight = Entity.createFlyweight(world);
+	}
 
-			end();
+	public EntitySubscription getSubscription() {
+		AspectSubscriptionManager sm = world.getManager(AspectSubscriptionManager.class);
+		return sm.get(aspectConfiguration);
+	}
+
+	@Override
+	public void inserted(ImmutableBag<Entity> entities) {
+		Object[] data = ((Bag<Entity>)entities).getData();
+		for (int i = 0, s = entities.size(); s > i; i++) {
+			inserted((Entity) data[i]);
 		}
 	}
-	
-	private void rebuildCompressedActives() {
-		
-		BitSet bs = activeIds;
-		int size = bs.cardinality();
-		actives.setSize(size);
-		actives.ensureCapacity(size);
-		EntityManager em = world.getEntityManager();
-		int[] activesArray = actives.getData();
-		for (int i = bs.nextSetBit(0), index = 0; i >= 0; i = bs.nextSetBit(i + 1)) {
-			activesArray[index++] = i;
-		}
-		
-		activesIsDirty = false;
-	}
-
-
-	/**
-	 * Called after the processing of entities ends.
-	 */
-	protected void end() {}
-	
-	/**
-	 * Any implementing entity system must implement this method and the logic
-	 * to process the given entities of the system.
-	 * 
-	 * @param entities
-	 *			the entities this system contains.
-	 */
-	protected abstract void processEntities(IntBag entities);
-	
-	/**
-	 * Check if the system should be processed.
-	 *
-	 * @return true if the system should be processed, false if not.
-	 */
-	@SuppressWarnings("static-method")
-	protected boolean checkProcessing() {
-		return true;
-	}
-
-	/**
-	 * Override to implement code that gets executed when systems are
-	 * initialized.
-	 */
-	protected void initialize() {}
 
 	/**
 	 * Called if the system has received a entity it is interested in, e.g
@@ -137,6 +72,14 @@ public abstract class EntitySystem implements EntityObserver {
 	 */
 	protected void inserted(Entity e) {}
 
+	@Override
+	public void removed(ImmutableBag<Entity> entities) {
+		Object[] data = ((Bag<Entity>)entities).getData();
+		for (int i = 0, s = entities.size(); s > i; i++) {
+			removed((Entity) data[i]);
+		}
+	}
+
 	/**
 	 * Called if a entity was removed from this system, e.g deleted or had one
 	 * of it's components removed.
@@ -146,159 +89,83 @@ public abstract class EntitySystem implements EntityObserver {
 	 */
 	protected void removed(Entity e) {}
 
-	/**
-	 * Returns true if the system is enabled.
-	 * 
-	 * @return {@code true} if enabled, otherwise false
-	 */
-	public boolean isEnabled() {
-		return enabled;
-	}
 
 	/**
-	 * Enabled systems are run during {@link #process()}.
-	 * <p>
-	 * Systems are enabled by default.
-	 * </p>
-	 * 
-	 * @param enabled
-	 *			system will not run when set to false
-	 */
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
-	
-	/**
-	 * A new unique component composition detected, check if this
-	 * system's aspect is interested in it.
-	 */
-	void processComponentIdentity(int id, BitSet componentBits) {
-		if (dummy)
-			return;
-
-		aspectCache.set(id, aspect.isInterested(componentBits));
-	}
-	
-	/**
-	 * Will check if the entity is of interest to this system.
+	 * Get all entities being processed by this system.
 	 *
-	 * @param e
-	 *			entity to check
+	 * @return a bag containing all active entities of the system
+	 *
+	 * @deprecated Retrieve the entities from the {@link com.artemis.EntitySubscription}
+	 * directly.
 	 */
-	protected final void check(Entity e) {
-		if(dummy)
-			return;
-		
-		EntityManager em = world.getEntityManager();
-		int id = e.getId();
-		boolean interested = aspectCache.get(em.getIdentity(e)) && em.isActive(id) && em.isEnabled(id);
-		boolean contains = activeIds.get(id);
-		
-		if (interested && !contains) {
-			insertToSystem(e);
-		} else if (!interested && contains) {
-			removeFromSystem(e);
+	@Deprecated
+	public Bag<Entity> getActives(Bag<Entity> fillBag) {
+		if (actives == null)
+			return fillBag;
+
+		int[] array = actives.getData();
+		for (int i = 0, s = actives.size(); s > i; i++) {
+			fillBag.add(world.getEntity(array[i]));
 		}
-	}
-	
-	/**
-	 * Removes the entity from this system.
-	 *
-	 * @param e
-	 *			the entity to remove
-	 */
-	private void removeFromSystem(Entity e) {
-		activeIds.clear(e.getId());
-		activesIsDirty = true;
-		
-		removed(e);
+
+		return fillBag;
 	}
 
 	/**
-	 * Inserts the entity into this system.
-	 *
-	 * @param e
-	 *			the entity to insert
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
 	 */
-	private void insertToSystem(Entity e) {
-		activeIds.set(e.getId());
-		activesIsDirty = true;
-		
-		inserted(e);
-	}
-	
-	/**
-	 * Call when an entity interesting to the system is added to the world.
-	 *
-	 * <p>
-	 * Checks if the system is interested in the added entity, if so, will
-	 * insert it in to the system.
-	 * </p>
-	 *
-	 * @param e
-	 *			the added entity
-	 */
-	@Override
-	public final void added(Entity e) {
-		check(e);
-	}
-	
-	@Override
-	public final void added(ImmutableBag<Entity> entities) {
-		Object[] data = ((Bag<Entity>)entities).getData();
-		for (int i = 0, s = entities.size(); s > i; i++) {
-			check((Entity)data[i]);
-		}
-	}
-	
-	@Override
-	public final void changed(ImmutableBag<Entity> entities) {
-		Object[] data = ((Bag<Entity>)entities).getData();
-		for (int i = 0, s = entities.size(); s > i; i++) {
-			check((Entity)data[i]);
-		}
-	}
-	
-	@Override
-	public final void deleted(ImmutableBag<Entity> entities) {
-		Object[] data = ((Bag<Entity>)entities).getData();
-		for (int i = 0, s = entities.size(); s > i; i++) {
-			Entity e = (Entity)data[i];
-			if (activeIds.get(e.getId()))
-				removeFromSystem(e);
-		}
-	}
+	@Deprecated
+	protected final void check(Entity e) {}
 
 	/**
-	 * Call when an entity interesting to the system has changed in the world.
-	 * <p>
-	 * Checks if the system is still interested after the entity has changed,
-	 * e.g a component was removed.
-	 * </p>
-	 *
-	 * @param e
-	 *			the changed entity
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
 	 */
-	@Override
-	public final void changed(Entity e) {
-		check(e);
-	}
+	@Deprecated
+	public final void added(Entity e) {}
 
 	/**
-	 * Call when an entity interesting to the system was deleted from the
-	 * world.
-	 * <p>
-	 * If the deleted entity is in this system, it will be removed.
-	 * </p>
-	 *
-	 * @param e
-	 *			the deleted entity
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
 	 */
-	@Override
-	public final void deleted(Entity e) {
-		if(activeIds.get(e.getId()))
-			removeFromSystem(e);
-	}
+	@Deprecated
+	public final void added(ImmutableBag<Entity> entities) {}
+
+	/**
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
+	 */
+	@Deprecated
+	public final void changed(ImmutableBag<Entity> entities) {}
+
+	/**
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
+	 */
+	@Deprecated
+	public final void deleted(ImmutableBag<Entity> entities) {}
+
+	/**
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
+	 */
+	@Deprecated
+	public final void changed(Entity e) {}
+
+	/**
+	 * This method no longer performs any operations due to entity subscriptions lists
+	 * being refactored into {@link com.artemis.AspectSubscriptionManager} and
+	 * {@link com.artemis.EntitySubscription}.
+	 */
+	@Deprecated
+	public final void deleted(Entity e) {}
 
 	/**
 	 * Call when an entity interesting to the system was disabled.
@@ -310,11 +177,8 @@ public abstract class EntitySystem implements EntityObserver {
 	 * @param e
 	 *			the disabled entity
 	 */
-	@Override
-	public final void disabled(Entity e) {
-		if(activeIds.get(e.getId()))
-			removeFromSystem(e);
-	}
+	@Deprecated
+	public final void disabled(Entity e) {}
 
 	/**
 	 * Call when an entity interesting to the system was (re)enabled.
@@ -325,76 +189,17 @@ public abstract class EntitySystem implements EntityObserver {
 	 * @param e
 	 *			the (re)enabled entity
 	 */
-	@Override
-	public final void enabled(Entity e) {
-		check(e);
-	}
-	
-	/**
-	 * Set the world this manager works on.
-	 *
-	 * @param world
-	 *			the world to set
-	 */
-	protected final void setWorld(World world) {
-		if (aspectConfiguration != null) {
-			aspect = aspectConfiguration.build(world);
-		} else {
-			dummy = true;
-		}
-		
-		this.world = world;
-	}
-
-	/**
-	 * Check if this system is passive.
-	 * <p>
-	 * A passive system will not process when {@link World#process()}
-	 * is called.
-	 * </p>
-	 *
-	 * @return {@code true} if the system is passive
-	 */
-	public boolean isPassive() {
-		return passive;
-	}
-
-	/**
-	 * Set if the system is passive or not.
-	 * <p>
-	 * A passive system will not process when {@link World#process()}
-	 * is called.
-	 * </p>
-	 *
-	 * @param passive
-	 *			{@code true} if passive, {@code false} if not
-	 */
-	protected void setPassive(boolean passive) {
-		this.passive = passive;
-	}
+	@Deprecated
+	public final void enabled(Entity e) {}
 
 	/**
 	 * Get all entities being processed by this system.
 	 *
 	 * @return a bag containing all active entities of the system
+	 * @deprecated This method allocates a new Bag each time, refer to {@link #getActives(com.artemis.utils.Bag)}
 	 */
 	@Deprecated
 	public ImmutableBag<Entity> getActives() {
-		if (activesIsDirty) // && world.isRebuildingIndexAllowed())
-			rebuildCompressedActives();
-
-		Bag<Entity> entities = new Bag<Entity>();
-
-		int[] array = actives.getData();
-		for (int i = 0, s = actives.size(); s > i; i++) {
-			entities.add(world.getEntity(array[i]));
-		}
-
-		return entities;
+		return getActives(new Bag<Entity>());
 	}
-
-	/**
-	 * see {@link World#dispose()}
-	 */
-	protected void dispose() {}
 }